add backgroundjobs

+++ src/Announcement.Application/AnnouncementService.cs
@@ -1,6 +1,10 @@
 ﻿using Announcement.DTOs;
+using AutoMapper;
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading;
 using System.Threading.Tasks;
 using Volo.Abp;
 using Volo.Abp.Application.Services;
@@ -44,7 +48,19 @@ namespace Announcement
         }
         public async Task DeleteAsync(Guid id)
         {
-           await _announcementRepository.DeleteAsync(id);
+            var existingAnnouncement = await _announcementRepository.FirstOrDefaultAsync(x => x.Id == id);
+            if (existingAnnouncement == null)
+            {
+                throw new UserFriendlyException("No announcements with this ID were found.");
+            }
+            await _announcementRepository.DeleteAsync(id);
         }
+
+        [RemoteService(IsEnabled = false)]
+        public async Task DeleteAnnouncementsOlderThanOneDay()
+        {
+            await _announcementRepository.DeleteDirectAsync(x => x.CreatedDate <= DateTime.Now.AddDays(-1));
+        }
+       
     }
 }
